+++ src/components/organisms/RouteEditingController/index.tsx
@@ -16,7 +16,7 @@ import { pagePaths } from "../../../consts/uriComponents";
 import styles from "./style.module.css";
 import { useAuthenticationInfoContext } from "../../../contexts/AuthenticationProvider";
 
-type RouteEditControllerProps = {
+type RouteEditingControllerProps = {
   isInsideMap: boolean;
   routeId: string;
   route: Route;
@@ -28,7 +28,7 @@ type RouteEditControllerProps = {
   setDrawingMode: React.Dispatch<React.SetStateAction<DrawingMode>>;
 };
 
-function RouteEditControllerDisplay(props: RouteEditControllerProps) {
+function RouteEditingControllerDisplay(props: RouteEditingControllerProps) {
   const history = useHistory();
   const { getIdToken } = useAuthenticationInfoContext();
 
@@ -130,7 +130,9 @@ function RouteEditControllerDisplay(props: RouteEditControllerProps) {
   );
 }
 
-export default function RouteEditController(props: RouteEditControllerProps) {
+export default function RouteEditingController(
+  props: RouteEditingControllerProps
+) {
   const divRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
@@ -149,11 +151,11 @@ export default function RouteEditController(props: RouteEditControllerProps) {
             className="leaflet-control leaflet-bar"
             style={{ width: 600 }}
           >
-            <RouteEditControllerDisplay {...props} />
+            <RouteEditingControllerDisplay {...props} />
           </div>
         </div>
       ) : (
-        <RouteEditControllerDisplay {...props} />
+        <RouteEditingControllerDisplay {...props} />
       )}
     </>
   );
